+++ processor/src/test/resources/expected/com/example/action/Arez_NoReportParametersModel.java
@@ -1,0 +1,61 @@
+package com.example.action;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.arez.Arez;
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.Disposable;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+public final class Arez_NoReportParametersModel extends NoReportParametersModel implements Disposable {
+  private static volatile long $$arez$$_nextId;
+
+  private final long $$arez$$_id;
+
+  private boolean $$arez$$_disposed;
+
+  @Nonnull
+  private final ArezContext $$arez$$_context;
+
+  public Arez_NoReportParametersModel() {
+    super();
+    this.$$arez$$_context = Arez.context();
+    this.$$arez$$_id = $$arez$$_nextId++;
+  }
+
+  final long $$arez$$_id() {
+    return $$arez$$_id;
+  }
+
+  String $$arez$$_name() {
+    return "NoReportParametersModel." + $$arez$$_id();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return $$arez$$_disposed;
+  }
+
+  @Override
+  public void dispose() {
+    if ( !isDisposed() ) {
+      $$arez$$_disposed = true;
+    }
+  }
+
+  @Override
+  public void doStuff(final long time, final float someOtherParameter) {
+    assert !$$arez$$_disposed;
+    try {
+      this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? $$arez$$_name() + ".doStuff" : null, true, () -> super.doStuff(time,someOtherParameter) );
+    } catch( final RuntimeException $$arez$$_e ) {
+      throw $$arez$$_e;
+    } catch( final Exception $$arez$$_e ) {
+      throw new IllegalStateException( $$arez$$_e );
+    } catch( final Error $$arez$$_e ) {
+      throw $$arez$$_e;
+    } catch( final Throwable $$arez$$_e ) {
+      throw new IllegalStateException( $$arez$$_e );
+    }
+  }
+}