+++ processor/src/test/resources/expected/com/example/tracked/Arez_NoReportParametersModel.java
@@ -1,0 +1,67 @@
+package com.example.tracked;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.arez.Arez;
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.Disposable;
+import org.realityforge.arez.Observer;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+public final class Arez_NoReportParametersModel extends NoReportParametersModel implements Disposable {
+  private static volatile long $$arez$$_nextId;
+
+  private final long $$arez$$_id;
+
+  private boolean $$arez$$_disposed;
+
+  @Nonnull
+  private final ArezContext $$arez$$_context;
+
+  @Nonnull
+  private final Observer $$arez$$_render;
+
+  public Arez_NoReportParametersModel() {
+    super();
+    this.$$arez$$_context = Arez.context();
+    this.$$arez$$_id = $$arez$$_nextId++;
+    this.$$arez$$_render = this.$$arez$$_context.tracker( this.$$arez$$_context.areNamesEnabled() ? $$arez$$_name() + ".render" : null, true, super::onRenderDepsUpdated );
+  }
+
+  final long $$arez$$_id() {
+    return $$arez$$_id;
+  }
+
+  String $$arez$$_name() {
+    return "NoReportParametersModel." + $$arez$$_id();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return $$arez$$_disposed;
+  }
+
+  @Override
+  public void dispose() {
+    if ( !isDisposed() ) {
+      $$arez$$_disposed = true;
+      $$arez$$_render.dispose();
+    }
+  }
+
+  @Override
+  public void render(final long time, final float someOtherParameter) {
+    assert !$$arez$$_disposed;
+    try {
+      this.$$arez$$_context.safeTrack( this.$$arez$$_render, () -> super.render(time,someOtherParameter) );
+    } catch( final RuntimeException $$arez$$_e ) {
+      throw $$arez$$_e;
+    } catch( final Exception $$arez$$_e ) {
+      throw new IllegalStateException( $$arez$$_e );
+    } catch( final Error $$arez$$_e ) {
+      throw $$arez$$_e;
+    } catch( final Throwable $$arez$$_e ) {
+      throw new IllegalStateException( $$arez$$_e );
+    }
+  }
+}